refactor(repo): move validation logic from builder into urlboxoptionsvalidation class
