+++ UrlboxSDK/Options/Builder/UrlboxOptionsBuilder.cs
@@ -7,59 +7,6 @@ public sealed class UrlboxOptionsBuilder
 {
     private readonly UrlboxOptions _options;
 
-    // ** Options that should not be applied if a given option is not set EG FullPage or UseS3 ** //
-
-    /// <summary>
-    /// A list of options that can only be used if full_page = true
-    /// </summary>
-    private static readonly string[] FullPageOptions =
-    {
-        nameof(UrlboxOptions.FullPageMode),
-        nameof(UrlboxOptions.ScrollIncrement),
-        nameof(UrlboxOptions.ScrollDelay),
-        nameof(UrlboxOptions.DetectFullHeight),
-        nameof(UrlboxOptions.MaxSectionHeight),
-        nameof(UrlboxOptions.FullWidth)
-    };
-
-    /// <summary>
-    /// A list of options that can only be used if use_s3 = true
-    /// </summary>
-    private static readonly string[] S3Options =
-    {
-        nameof(UrlboxOptions.S3Bucket),
-        nameof(UrlboxOptions.S3Path),
-        nameof(UrlboxOptions.S3Endpoint),
-        nameof(UrlboxOptions.S3Region),
-        nameof(UrlboxOptions.S3Storageclass),
-        nameof(UrlboxOptions.CdnHost),
-    };
-
-    // Define PDF-specific options as a static readonly field
-    private static readonly string[] PdfOptions =
-    {
-        nameof(UrlboxOptions.PdfPageSize),
-        nameof(UrlboxOptions.PdfPageRange),
-        nameof(UrlboxOptions.PdfPageWidth),
-        nameof(UrlboxOptions.PdfPageHeight),
-        nameof(UrlboxOptions.PdfMargin),
-        nameof(UrlboxOptions.PdfMarginTop),
-        nameof(UrlboxOptions.PdfMarginRight),
-        nameof(UrlboxOptions.PdfMarginBottom),
-        nameof(UrlboxOptions.PdfMarginLeft),
-        nameof(UrlboxOptions.PdfAutoCrop),
-        nameof(UrlboxOptions.PdfScale),
-        nameof(UrlboxOptions.PdfOrientation),
-        nameof(UrlboxOptions.PdfBackground),
-        nameof(UrlboxOptions.DisableLigatures),
-        nameof(UrlboxOptions.Media),
-        nameof(UrlboxOptions.Readable),
-        nameof(UrlboxOptions.PdfShowHeader),
-        nameof(UrlboxOptions.PdfHeader),
-        nameof(UrlboxOptions.PdfShowFooter),
-        nameof(UrlboxOptions.PdfFooter)
-    };
-
     /// <summary>
     /// Constructor
     /// </summary>
@@ -79,95 +26,7 @@ public sealed class UrlboxOptionsBuilder
     /// <returns></returns>
     public UrlboxOptions Build()
     {
-        return Validate(_options);
-    }
-
-    private static UrlboxOptions Validate(UrlboxOptions options)
-    {
-        ValidateScreenshotOptions(options);
-        ValidatePdfOptions(options);
-        ValidateFullPageOptions(options);
-        ValidateS3Options(options);
-        return options;
-    }
-
-    private static UrlboxOptions ValidateScreenshotOptions(UrlboxOptions options)
-    {
-        var thumbSizes = options.ThumbWidth != null || options.ThumbHeight != null;
-        bool hasImgFit = options.ImgFit != null && Enum.IsDefined(typeof(ImgFit), options.ImgFit);
-        bool hasImgPosition = options.ImgPosition != null && Enum.IsDefined(typeof(ImgPosition), options.ImgPosition);
-        var imgFitIsCoverOrContain = options.ImgFit == UrlboxSDK.Options.Resource.ImgFit.Cover || options.ImgFit == UrlboxSDK.Options.Resource.ImgFit.Contain;
-
-        if (!thumbSizes && hasImgFit)
-        {
-            throw new ArgumentException("Invalid Configuration: Image Fit is included despite ThumbWidth nor ThumbHeight being set.");
-        }
-
-        if (!hasImgFit && hasImgPosition)
-        {
-            throw new ArgumentException("Invalid Configuration: Image Position is included despite Image Fit not being set.");
-        }
-
-        if (hasImgFit && hasImgPosition && !imgFitIsCoverOrContain)
-        {
-            throw new ArgumentException("Invalid Configuration: Image Position is included despite Image Fit not being set to 'cover' or 'contain'.");
-        }
-
-        return options;
-    }
-
-    private static UrlboxOptions ValidateFullPageOptions(UrlboxOptions options)
-    {
-        bool isNotFullPage = !options.FullPage.HasValue || (options.FullPage.HasValue && options.FullPage.Value.Bool != true);
-        bool hasFullPageOptions = HasOptionsInCategory(FullPageOptions, options);
-        if (
-            isNotFullPage && hasFullPageOptions
-        )
-        {
-            throw new ArgumentException("Invalid configuration: Full-page options are included despite 'FullPage' being set to false.");
-        }
-        return options;
-    }
-
-    private static UrlboxOptions ValidateS3Options(UrlboxOptions options)
-    {
-        bool isNotUsingS3 = !options.UseS3.HasValue || (options.UseS3.HasValue && options.UseS3.Value.Bool != true);
-        bool hasS3Options = HasOptionsInCategory(S3Options, options);
-        if (isNotUsingS3 && hasS3Options)
-        {
-            throw new ArgumentException("Invalid configuration: S3 options are included despite 'UseS3' being set to false.");
-        }
-        return options;
-    }
-
-    private static UrlboxOptions ValidatePdfOptions(UrlboxOptions options)
-    {
-        bool isNotUsingPdf = options.Format != UrlboxSDK.Options.Resource.Format.Pdf;
-        bool hasPdfOptions = HasOptionsInCategory(PdfOptions, options);
-        if (isNotUsingPdf && hasPdfOptions)
-        {
-            throw new ArgumentException("One or more PDF-specific options are only valid for the PDF format.");
-        }
-
-        return options;
-    }
-
-    /// <summary>
-    /// Determines if any properties in the specified category are set in the given options.
-    /// </summary>
-    /// <param name="category">Array of property names to check within the options.</param>
-    /// <param name="options">The options object to inspect.</param>
-    /// <returns>True if any property in the category is set; otherwise, false.</returns>
-    private static bool HasOptionsInCategory(string[] category, UrlboxOptions options)
-    {
-        return category
-         .Any(propertyName =>
-         {
-             var property = options.GetType().GetProperty(propertyName);
-             if (property == null) return false;
-             var value = property.GetValue(options);
-             return UrlboxOptionsValidation.IsNullOption(value);
-         });
+        return UrlboxOptionsValidation.Validate(_options);
     }
 
     public UrlboxOptionsBuilder WebhookUrl(string webhookUrl)
